Proved the following invariant on continuation hole semantics:
There is no "hole" in the continuation stack apart from the top.

this invariant is preserved by the cred relation and should be useful in
the sred_to_cred proof for the following case.

(sred (Default
         (fst (apply_conts kappa (apply_return result, env0))
          :: ts..[subst_of_env (last kappa env0)])
         tj0.[subst_of_env (last kappa env0)]
         tc0.[subst_of_env (last kappa env0)])
      (Default
         (fst (apply_state_aux x) :: ts..[subst_of_env (last kappa env0)])
         tj0.[subst_of_env (last kappa env0)]
         tc0.[subst_of_env (last kappa env0)]))

with (mode_cont (kappa ++ [CDefault Hole None ts tj0 tc0]) env0 result)
equivalent to the left hand side of sred.